data_download/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "wsfr-download"
authors = [{ name = "DrivenData", email = "info@drivendata.org" }]
license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.10"
dependencies = [
  "loguru",
  "netcdf4",
  "numpy",
  "pandas",
  "pydantic",
  "pyyaml",
  "requests",
  "tqdm",
  "typer",
]

[project.urls]
Home = "https://github.com/drivendataorg/water-supply-forecast-rodeo-runtime"

## TOOLS

[tool.black]
line-length = 99
include = '\.pyi?$'

[tool.ruff]
line-length = 99
select = [
  "E", # Pyflakes
  "F", # Pycodestyle
  "I", # isort
]
src = ["wsfr_download"]
exclude = []
unfixable = ["F"]

[tool.ruff.isort]
known-first-party = ["wsfr_download"]
force-sort-within-sections = true
